requirements.txt
websocket-client==1.6.1
confluent-kafka==2.2.0
clickhouse-driver==0.2.6
numpy==1.21.6
pandas==2.0.3 
retry>=0.9.2
